Moves appropriate Pages into Pages folder and adds React Router Main Layout to style correctly
