--- client/src/components/About.jsx
@@ -1,20 +1,0 @@
-//Import necessary functionalities
-import React from 'react';
-
-const About = () => {
-
-    console.log("This is the about page!");
-
-    return (
-        <div>
-            <h1>About Page</h1>
-
-            {/* Add in actual information as well as relevant pictures! */}
-            <p>Blurb about the pages!</p>
-            <p>Info about Techtonica, Tati, and Daisy!</p>
-            
-        </div>
-    )
-}
-
-export default About;
