--- client/src/components/Examples.jsx
@@ -1,19 +1,0 @@
-//Import necessary functionalities
-import React from 'react';
-
-const Examples = () => {
-
-    return (
-        <div>
-            <h1>Example Training Plans</h1>
-
-            <h3>Currently Offered Plans</h3>
-            <p>Information about Middle and how to complete it goes here</p>
-            
-            <h3>Other Plans</h3>
-            <p>Details about plans such as Touch and Vacuum Monster go here!</p>
-        </div>
-    )
-}
-
-export default Examples;
