+++ client/src/pages/Examples.jsx
@@ -1,0 +1,22 @@
+//Import necessary functionalities
+import React from 'react';
+import MainLayout from '../layout/MainLayout';
+
+const Examples = () => {
+
+    return (
+        <MainLayout>
+            <div>
+                <h1>Example Training Plans</h1>
+
+                <h3>Currently Offered Plans</h3>
+                <p>Information about Middle and how to complete it goes here</p>
+                
+                <h3>Other Plans</h3>
+                <p>Details about plans such as Touch and Vacuum Monster go here!</p>
+            </div>
+        </MainLayout>
+    )
+}
+
+export default Examples;
